[FE] refactor: 마이페이지 다이어리 컴포넌트 파일명 변경
