+++ FE/src/components/mypage/details/myInfo/MyInfo.tsx
