+++ FE/src/pages/MyPage.tsx
@@ -3,9 +3,9 @@ import styles from "@styles/mypage/MyPage.module.scss";
 import Profile from "@components/common/Profile";
 import { profileInfo } from "./commonPage";
 import MyPageTab from "@components/mypage/MyPageTab";
-import MyPageDiary from "@components/mypage/details/diary/MyPageDiary";
+import MyPageDiary from "@components/mypage/details/diary/Diary";
 import MyPageFrame from "@components/mypage/MyPageFrame";
-import MyPageInfo from "@components/mypage/details/myInfo/MyPageInfo";
+import MyPageInfo from "@components/mypage/details/myInfo/MyInfo";
 import Following from "@components/mypage/details/friend/Following";
 import Follower from "@components/mypage/details/friend/Follower";
 import Search from "@components/mypage/details/friend/Search";
